+++ functions/Lyrics.py
