+++ routes/templates/Canciones.html
@@ -1,0 +1,67 @@
+<!DOCTYPE html>
+<html lang="es">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Lista de Canciones</title>
+    <style>
+        body {
+            color: black; /* Color del texto */
+            font-family: 'Arial', sans-serif;
+            margin: 0;
+            padding: 0;
+            display: flex;
+            flex-direction: column;
+            align-items: center;
+            justify-content: center;
+            height: 100vh;
+            text-align: center;
+        }
+
+        .artist-image {
+            border-radius: 50%;
+            width: 100px;
+            height: 100px;
+            margin-bottom: 10px;
+        }
+
+        h1 {
+            padding: 10px;
+            animation: neon 1.5s ease-in-out infinite alternate;
+        }
+
+        @keyframes neon {
+            from {
+                text-shadow: 0 0 10px #fff, 0 0 20px #fff, 0 0 30px #fff;
+            }
+            to {
+                text-shadow: 0 0 10px #f06, 0 0 20px #f06, 0 0 30px #f06;
+            }
+        }
+
+        ul {
+            list-style-type: none;
+            padding: 0;
+        }
+
+        li {
+            margin: 5px 0;
+        }
+
+        .average-info {
+            font-size: 18px;
+            margin-top: 20px;
+            color: black; /* Color del texto de la información del promedio */
+        }
+    </style>
+</head>
+<body>
+    <img src="{{ image_url }}" alt="Imagen del artista" class="artist-image">
+    <h1>Canciones del álbum '{{ album_name }}' de '{{ artist_name }}'</h1>
+    <ul>
+        {% for track in tracks %}
+            <li>{{ track['name'] }}</li>
+        {% endfor %}
+    </ul>
+</body>
+</html>
